pulled many API implementations to separate files
