+++ sympy/simplify/__init__.py
@@ -6,7 +6,7 @@ the expression (x+x)**2 will be converted into 4*x**2
 """
 from .simplify import (simplify, hypersimp, hypersimilar,
     logcombine, separatevars, posify, besselsimp, kroneckersimp,
-    signsimp, bottom_up, nsimplify)
+    signsimp, bottom_up, nsimplify, dotprodsimp)
 
 from .fu import FU, fu
 
@@ -35,7 +35,7 @@ from .ratsimp import ratsimp, ratsimpmodprime
 __all__ = [
     'simplify', 'hypersimp', 'hypersimilar', 'logcombine', 'separatevars',
     'posify', 'besselsimp', 'kroneckersimp', 'signsimp', 'bottom_up',
-    'nsimplify',
+    'nsimplify', 'dotprodsimp',
 
     'FU', 'fu',
